Camera script finished.
- I have finished working on the camera script; it can now move from P1, to P2 or Intermission. The values at this moment are hardcoded, but there is a function that takes arguments to set the new position for the camera to move to.

SetNewCameraPosition(Vector3 newPosition), is used to set new position for the camera, it takes a Vector3 because it is happening in 3D world space (this is how Unity works), but the 'z' axis should always be set to '-10' (camera layer).

SetNewCameraSize is used to change the ortographic camera size, it can be used to resize the camera when for example in the Intermission state when the camera is showing both players. It is currently not chaning the FOV of the camera; this will be implemented later.

- I will now start working on the Turn Based System.