Started work on the countdown timer
Made additions to the GameManager. Eventually, the turn will change after 30 seconds with the remaining time being visible on screen.

Modified the block breaking script to put another variable at the top of the script. This means I do not have to search through the whole script to change the break chance.

Also installed TextMesh Pro. This can be used to make the text in the UI and menu look nicer (not fuzzy)