Further work on the countdown timer
There is a problem with the IsCountingDown bool. For some reason, Unity stops responding when I try to play the game. I have commented out all of the code relating to the bool and the game is playable again, so work on the GameManager can continue without it. The bool is to stop and reset the countdown while neither player is taking their turn.