Implemented new Turn System.
I have implemented new turn system, rather than using enums and IEnumerators for the turn system, I have now changed it so that it does not use them.

The game in the start of the round finds all players, and adds them to an array. Int playerNumber keeps track of which player's turn it is (1 = P1, 2 = P2, etc.), and the player scripts all read from the GameManger to determine if its their turn or not.

Intermission is now a separate function, so now we can add a lot of functionality to it such as the powerup drops, and other functions.  I have also used this system to limit the player's ability to shoot; while previosuly players were still able to shoot in the "Intermission" stage of the game, now they can only shoot when its their turn.

Other changes include getting rid of some variables that are no longer needed, hiding variables from editor that don't need to be edited and are working properly and more general optimisation/cleaning.