Aim locking system.
I have added a new feature where once player taps on the screen to select an angle, the angle is locked in and the player needs to press space to carry out the shot.

This is the fundamentals for additional mechanics that we will add later on to the game. The next step for me is to create the trajectory that will show the player where their projectile will go when they have selected the aim angle.