Started work on the time-based mechanics
Added in some time variables. Had issues getting turn timers to work properly, but they just about worked with the damage multiplier. Accidentally broke the turn counter but I later fixed that. If the camera pans to the wrong player, the countdown timer has messed up again. Increasing timeLimit and currentTimeLimit should fix that. Floats are a pain.
Also changed the input to click + space to two clicks. This might change to a single click in future.