Re-imported assets & Created alpha scene
I ahve created the alpha scene for our prototype where I have re-imported the assets and added a simple tower for testing. This is the scene we will work on from now until we get to the beta of the game.

Next week we will work on the modular buildings so that we can start puttting levels together and test the game.