UI Changes and Level Design
I went and changed the Main menu to no longer include a "Quit button"
and shuffled the other elements to fit better. I also changed the title
to it's final form and made sure the sprites used were not stretched.

I also constructed a small new structure to start to form ideas on
future and more unique structures as part of the level design, this is
to be shown and discussed on the 20th of April during a group meeting.